requirements.txt
Django>=4.2.0
Pillow>=10.0.0
requests>=2.31.0
python-decouple>=3.8
whitenoise>=6.5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-phonenumber-field>=7.1.0
phonenumbers>=8.13.0
twilio>=8.10.0
python-dotenv>=1.0.0
reportlab>=4.0.0
google-generativeai>=0.3.0
django-import-export>=4.3.14

# Real-time WebSocket notifications
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
redis>=5.0.0
